+++ crates/texlab/src/util/to_proto.rs
@@ -387,6 +387,7 @@ pub fn symbol_kind(value: symbols::SymbolKind) -> lsp_types::SymbolKind {
         symbols::SymbolKind::Field => lsp_types::SymbolKind::FIELD,
         symbols::SymbolKind::Environment => lsp_types::SymbolKind::FUNCTION,
         symbols::SymbolKind::CommandDefinition => lsp_types::SymbolKind::KEY,
+        symbols::SymbolKind::BeamerFrame => lsp_types::SymbolKind::CLASS,
     }
 }