+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add document symbols for beamer frames ([[#1498](https://github.com/latex-lsp/texlab/issues/1498)])
+
 ### Fixed
 
 - Don't allow line breaks in URLs ([#1488](https://github.com/latex-lsp/texlab/issues/1488))